+++ docs/Changelog.md
@@ -45,6 +45,7 @@
 * [#5197](https://github.com/TouK/nussknacker/pull/5197) Improved accessing fields in records in expressions - fields can now be statically accessed using indexing
 * [#5312](https://github.com/TouK/nussknacker/pull/5312) Collect component clears context variables
 * [#5313](https://github.com/TouK/nussknacker/pull/5313) Added CountWhen and Average aggregations
+* [#5522](https://github.com/TouK/nussknacker/pull/5522) Scenario status caching more often
 
 1.12.5 (1 Dec 2023)
 ------------------------